Enable AutoProfiling for golang component
