+++ dispatch/go.mod
@@ -1,0 +1,8 @@
+module github.com/instana/robot-shop/dispatch
+
+go 1.14
+
+require (
+	github.com/instana/go-sensor v1.13.1 // indirect
+	github.com/streadway/amqp v1.0.0 // indirect
+)
